+++ contracts/SafeMath.sol
@@ -37,11 +37,11 @@ library SafeMath {
     return a < b ? a : b;
   }
 
-  function max256(uint256 a, uint256 b) internal constant returns (uint256) {
+  function max256(uint a, uint b) internal constant returns (uint) {
     return a >= b ? a : b;
   }
 
-  function min256(uint256 a, uint256 b) internal constant returns (uint256) {
+  function min256(uint a, uint b) internal constant returns (uint) {
     return a < b ? a : b;
   }
